+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.0.6 - 2022-04-25
+==================
+### Fixes
+ - Better name "kptncook-today" for the command fetching the 3 kptncook recipes for today
+    - #14 issue by @ephes
+ - Fixed __all__ exports
+    - #14 issue by @ephes
+
 0.0.5 - 2022-04-20
 ==================
 ### Refactoring